library name template
